173 - check for last wf run only
