+++ api/src/services/dataset.js
@@ -5,6 +5,7 @@ const { PrismaClient } = require('@prisma/client');
 const config = require('config');
 // const _ = require('lodash/fp');
 
+const dayjs = require('dayjs');
 const wfService = require('./workflow');
 const userService = require('./user');
 const { log_axios_error } = require('../utils');
@@ -444,10 +445,60 @@ async function add_files({ dataset_id, data }) {
   });
 }
 
+/**
+ * Get workflows for a dataset
+ * @param dataset_id    the id of the dataset whose workflows are to be retrieved
+ * @param last_run_only if true, only the last run of each workflow is returned
+ * @param statuses      an array of workflow statuses to filter retrieved workflows by
+ * @returns Array of workflows
+ */
+async function get_workflows({ dataset_id, last_run_only = false, statuses = [] }) {
+  const workflows = await prisma.workflow.findMany({
+    where: {
+      dataset_id,
+    },
+  });
+  const workflow_ids = workflows.map((w) => w.id);
+
+  const wf_promises = workflow_ids.map((id) => wfService.getOne(id));
+
+  const retrievedWorkflowsResponses = await Promise.all(wf_promises);
+  let retrievedWorkflows = retrievedWorkflowsResponses.map((e) => e.data);
+
+  if (last_run_only) {
+    // filter last successful runs
+    let workflow_names = retrievedWorkflows
+      .map((wf) => wf.name);
+      // get distinct workflow names
+    workflow_names = workflow_names
+      .filter((name, i) => workflow_names.indexOf(name) === i);
+
+    // group workflows by distinct workflow names
+    const workflows_grouped_by_name = workflow_names.map((wf_name) => (
+      retrievedWorkflows
+        .filter((wf) => wf.name === wf_name)));
+
+    retrievedWorkflows = workflows_grouped_by_name.map(
+      (wfs_grouped_by_name) => wfs_grouped_by_name.reduce(
+        (acc, curr) => (dayjs(acc.updated_at).diff(dayjs(curr.updated_at)) >= 0 ? acc : curr),
+      ),
+    );
+
+    // console.log('retrievedWorkflows');
+    // console.dir(retrievedWorkflows, { depth: null });
+  }
+
+  // filter by status
+  if ((statuses || []).length > 0) {
+    retrievedWorkflows = retrievedWorkflows.filter((wf) => statuses.includes(wf.status));
+  }
+}
+
 module.exports = {
   soft_delete,
   get_dataset,
   create_workflow,
+  get_workflows,
   create_filetree,
   has_dataset_assoc,
   files_ls,
